feat: added Docker configuration
