+++ requirements.txt
@@ -1,4 +1,3 @@
 openai-agents==0.0.14
-fastapi>=0.103.1
-uvicorn>=0.23.2
+fastapi[standard]>=0.113.0
 strenum>=0.4.15
